requirements.txt
anyio==4.8.0
asgiref==3.8.1
boto3==1.36.11
botocore==1.36.11
certifi==2025.1.31
Django==5.1.5
django-cors-headers==4.6.0
django-lifecycle==1.2.4
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
psycopg2==2.9.10
psycopg2-binary==2.9.10
dj-database-url==2.1.0
sqlparse==0.5.3
dj-static==0.0.6
static==1.1.1
static3==0.7.0